backend/requirements.txt
fastapi
uvicorn[standard]
sentence-transformers
faiss-cpu
openai
pandas
tqdm
scikit-learn
transformers
datasets
numpy
python-multipart